requirements.txt
asgiref==3.3.4
autopep8==1.5.4
Django==3.1.5
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.9.0
gunicorn==20.0.4
mccabe==0.6.1
model-bakery==1.2.1
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
